feat: create and use isUserAdmin method in services

+++ src/user-role/user-role.service.ts
@@ -1,35 +1,56 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
 import { CreateUserRoleDto } from './dto/create-user-role.dto';
 import { UpdateUserRoleDto } from './dto/update-user-role.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { UserRole } from './entities/user-role.entity';
 import { Repository } from 'typeorm';
+import { User } from '../users/entities/user.entity';
+import { Role } from '../setup/enums/role.enum';
+import { RoleEntity } from '../role/entities/role.entity';
+import { RoleService } from '../role/role.service';
 
 @Injectable()
 export class UserRoleService {
   constructor(
     @InjectRepository(UserRole)
-    private userRoleRepository: Repository<UserRole>
+    private userRoleRepository: Repository<UserRole>, 
+    private readonly roleService: RoleService
   ){}
 
-  create(createUserRoleDto: CreateUserRoleDto) {
+  create(userReq: User, createUserRoleDto: CreateUserRoleDto) {
+    const isAdmin = this.roleService.isUserAdmin(userReq);
+
+    if (!isAdmin) {
+      throw new ForbiddenException('Você não tem permissão para adicionar uma nova regra!')
+    }
     const userRole = this.userRoleRepository.create(createUserRoleDto);
     return this.userRoleRepository.save(userRole);
   }
 
-  findAll() {
+  findAll(userReq: User) {
+    const isAdmin = this.roleService.isUserAdmin(userReq);
+
+    if (!isAdmin) {
+      throw new ForbiddenException()
+    }
     return this.userRoleRepository.find();
   }
 
-  findOne(roleId: number, userId: number) {
+  findOne(userReq: User,roleId: number, userId: number) {
+    const isAdmin = this.roleService.isUserAdmin(userReq);
+
+    if (!isAdmin) {
+      throw new ForbiddenException()
+    }
     return this.userRoleRepository.findOneBy({ roleId: roleId, userId });
   }
 
-  update(roleId: number, userId: number, updateUserRoleDto: UpdateUserRoleDto) {
-    return this.userRoleRepository.update({ roleId, userId }, updateUserRoleDto);
-  }
+  remove(userReq: User, roleId: number, userId: number) {
+    const isAdmin = this.roleService.isUserAdmin(userReq);
 
-  remove(roleId: number, userId: number) {
+    if (!isAdmin) {
+      throw new ForbiddenException();
+    }
     return this.userRoleRepository.delete({ roleId, userId });
   }
 }
